[refactor] 회원가입 1단계 로직 수정
