+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Email/Register/Dto/RegResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.LifeMaster_BE.UserManager.Email.Register.Dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class RegResponseDto {
+    private String regId;
+}
