+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Email/Register/Dto/RegisterDto.java
@@ -1,7 +1,6 @@
-package com.example.LifeMaster_BE.UserManager.Email.Register;
+package com.example.LifeMaster_BE.UserManager.Email.Register.Dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.*;
