+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Email/Register/RegisterController.java
@@ -1,12 +1,17 @@
 package com.example.LifeMaster_BE.UserManager.Email.Register;
 
+import com.example.LifeMaster_BE.UserManager.Email.Register.Dto.RegResponseDto;
+import com.example.LifeMaster_BE.UserManager.Email.Register.Dto.RegisterDto;
+import com.example.LifeMaster_BE.UserManager.Email.Register.Dto.RegisterWithNicknameDto;
 import com.example.LifeMaster_BE.UserManager.Member.MemberEntity;
 import com.example.LifeMaster_BE.UserManager.Peristalsis.OAuthUsersRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 @Slf4j
 @RestController
@@ -18,18 +23,19 @@ public class RegisterController {
     private final OAuthUsersRepository OAuthUsersRepository;
 
     @PostMapping
-    public ResponseEntity<Object> register(@RequestBody RegisterDto registerDto) {
-        String password = registerDto.getPassword();
-        String passwordConfirm = registerDto.getPasswordConfirm();
+    public ResponseEntity<RegResponseDto> register(@RequestBody RegisterDto registerDto) {
         String email = registerDto.getEmail();
 
         // 연동 로그인 이메일 사용 방지
         if (OAuthUsersRepository.existsByEmail(email)) {
-            return ResponseEntity.badRequest().body("이미 해당 계정은 연동 계정입니다. 연동 로그인을 이용하세요.");
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "이미 해당 계정은 연동 계정입니다. 연동 로그인을 이용하세요."
+            );
         }
 
-        MemberEntity memberEntity = registerService.registerMember(email, password, passwordConfirm);
-        return ResponseEntity.ok(memberEntity);
+        RegResponseDto regResponse = registerService.registerMember(registerDto);
+        return ResponseEntity.ok(regResponse);
     }
 
     @PostMapping("/nickname")
